fix error handling
