+++ ledger/acctupdates.go
@@ -32,7 +32,6 @@ import (
 
 	"github.com/algorand/go-deadlock"
 
-	"github.com/algorand/go-algorand/agreement"
 	"github.com/algorand/go-algorand/config"
 	"github.com/algorand/go-algorand/crypto"
 	"github.com/algorand/go-algorand/crypto/merkletrie"
@@ -212,6 +211,16 @@ type deferedCommit struct {
 	offset   uint64
 	dbRound  basics.Round
 	lookback basics.Round
+}
+
+// RoundOffsetError is an error for when requested round is behind earliest stored db entry
+type RoundOffsetError struct {
+	Round   basics.Round
+	DbRound basics.Round
+}
+
+func (e *RoundOffsetError) Error() string {
+	return fmt.Sprintf("round %d before dbRound %d", e.Round, e.DbRound)
 }
 
 // initialize initializes the accountUpdates structure
@@ -1431,7 +1440,7 @@ func (au *accountUpdates) accountsCreateCatchpointLabel(committedRound basics.Ro
 // roundOffset calculates the offset of the given round compared to the current dbRound. Requires that the lock would be taken.
 func (au *accountUpdates) roundOffset(rnd basics.Round) (offset uint64, err error) {
 	if rnd < au.dbRound {
-		err = &agreement.RoundOffsetError{
+		err = &RoundOffsetError{
 			Round:   rnd,
 			DbRound: au.dbRound,
 		}
